pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "reviewbot"
dynamic = ["version"]
description = "Git-native AI code review CLI with schema-validated LLM output."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12.0",
    "rich>=13.7.0",
    "pydantic>=2",
    "httpx>=0.27.0",
    "groq>=0.11.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest"]

[project.scripts]
reviewbot = "reviewbot.cli:main"

[tool.setuptools]
packages = ["reviewbot"]

[tool.setuptools.dynamic]
version = {attr = "reviewbot.__version__"}
